Description: -Fix endianess issue on HP-UX. (Thanks Nick Baronian for testing it).
-Added "srcport", "dstport", "same_src_port" and "same_dst_port" as options to
the rules.
(Thanks Michael Starks for the suggestion).

-Added "maild.full_subject" to internal_options.conf to allow more
verbose subjects on the e-mail notifications (set it to "1" to include
the rule description).
(Thanks Michael Starks for the suggestion).

Reviewed by: dcid
Bug:

+++ src/shared/read-agents.c
@@ -35,7 +35,7 @@ void free_agents(char **agent_list)
 
 
 /* Delete syscheck db */ 
-int delete_syscheck(char *sk_name, char *sk_ip)
+int delete_syscheck(char *sk_name, char *sk_ip, int full_delete)
 {
     FILE *fp;
     char tmp_file[513];
@@ -51,7 +51,9 @@ int delete_syscheck(char *sk_name, char *sk_ip)
     fp = fopen(tmp_file, "w");
     if(fp)
         fclose(fp);
-    unlink(tmp_file);
+
+    if(full_delete)    
+        unlink(tmp_file);
 
 
     /* Deleting cpt files */
@@ -75,7 +77,8 @@ int delete_syscheck(char *sk_name, char *sk_ip)
     fp = fopen(tmp_file, "w");
     if(fp)
         fclose(fp);
-    unlink(tmp_file);
+    if(full_delete)
+        unlink(tmp_file);
 
 
     /* Deleting cpt files */
@@ -120,7 +123,7 @@ int delete_agentinfo(char *name)
 
 
     /* Deleting syscheck */
-    delete_syscheck(sk_name, sk_ip);
+    delete_syscheck(sk_name, sk_ip, 1);
     
     return(1);
 }
